+++ code/gsc.cpp
@@ -159,6 +159,7 @@ scr_function_t scriptFunctions[] =
 	{"getEntityCount", gsc_level_getentitycount, 0},
 	{"getMovers", gsc_level_getmovers, 0},
 	{"getPlayersInRange", gsc_level_getplayersinrange, 0},
+	{"getClosestPlayerInRange", gsc_level_getclosestplayerinrange, 0},
 	{"getNumberOfStaticModels", gsc_level_getnumberofstaticmodels, 0},
 	{"getSavePersist", gsc_level_getsavepersist, 0},
 	{"getStaticModelName", gsc_level_getstaticmodelname, 0},